Add query statuses and move scheduling worker
